refactor: EnginePackage系の型をdomain/enginePackage.tsに移動
vueコンポーネントがbackend/electronを直接参照しないよう、
EnginePackageBase/CurrentInfo/LatestInfoをsrc/domain/enginePackage.tsに移動し、
各ファイルのインポートを更新する。

ref: https://github.com/VOICEVOX/voicevox/issues/2927
